Remove deprecated OAuth1 and update to current GCM API.
Also switched build system to gradle, so it can compile nicely in AndroidStudio.

Various refactorings - using messages and more robust setup activity.

For authentication we use audience tokens - the main limiatation is that on Gingerbread
it will require the install of PlayServices, if it is not already installed. On ICS plus
it will use it if present, or fall back to platform AccountManager.